Refactor formatter classes for `data` and `view<data>` classes
Move formatters to header files dedicated to formatted classes. This also required extra boilerplate to make it possible for `view<data>` formatter to reuse `data` formatter.
