+++ libvast/vast/fmt_integration.hpp
@@ -16,12 +16,10 @@
 #include "vast/fwd.hpp"
 
 #include "vast/aliases.hpp"
-#include "vast/data.hpp"
 #include "vast/detail/escapers.hpp"
-#include "vast/view.hpp"
+#include "vast/time.hpp"
 
 #include <chrono>
-#include <string_view>
 #include <type_traits>
 
 namespace vast::detail {
@@ -288,322 +286,5 @@ struct formatter<vast::pattern> {
     return format_to(ctx.out(), "/{}/", p.string());
   }
 };
-
-template <>
-struct formatter<vast::view<vast::pattern>> : formatter<vast::pattern> {};
-
-// Specialization which implements formatting of `<data,data>` pair.
-template <>
-struct formatter<vast::map::value_type> : vast::detail::empty_formatter_base {
-  template <class ValueType, class FormatContext>
-  auto format(const ValueType& v, FormatContext& ctx) {
-    return fmt::format_to(ctx.out(), "{} -> {}", v.first, v.second);
-  }
-};
-
-template <>
-struct formatter<std::pair<vast::view<vast::data>, vast::view<vast::data>>>
-  : formatter<vast::map::value_type> {};
-
-// Specialization which implements formatting of `<string,data>` pair.
-template <>
-struct formatter<vast::record::value_type>
-  : vast::detail::empty_formatter_base {
-  template <class ValueType, class FormatContext>
-  auto format(const ValueType& v, FormatContext& ctx) {
-    return fmt::format_to(ctx.out(), "{}: {}", v.first, v.second);
-  }
-};
-
-template <>
-struct formatter<std::pair<std::string_view, vast::view<vast::data>>>
-  : formatter<vast::record::value_type> {};
-
-/// Definition of fmt-formatting rules for vast::data.
-template <>
-struct formatter<vast::data> : public vast::detail::vast_formatter_base {
-  using ascii_escape_string
-    = vast::detail::escaped_string_view<vast::detail::print_escaper_functor>;
-  using json_escape_string
-    = vast::detail::escaped_string_view<vast::detail::json_escaper_functor>;
-
-  template <typename Output>
-  struct ascii_visitor {
-    Output out_;
-
-    ascii_visitor(Output out) : out_{std::move(out)} {
-    }
-
-    // bool operator()(const vast::view<vast::data>& x) {
-    //   return caf::visit(*this, x);
-    // }
-    auto operator()(caf::none_t) {
-      return format_to(out_, "nil");
-    }
-    auto operator()(bool b) {
-      return format_to(out_, b ? "T" : "F");
-    }
-    auto operator()(vast::duration d) {
-      return format_to(out_, "{}",
-                       vast::detail::fmt_wrapped<vast::duration>{d});
-    }
-    auto operator()(const vast::time& t) {
-      return format_to(out_, "{}", vast::detail::fmt_wrapped<vast::time>{t});
-    }
-    auto operator()(const std::string& s) {
-      return (*this)(ascii_escape_string{s});
-    }
-    auto operator()(const std::string_view& s) {
-      return (*this)(ascii_escape_string{s});
-    }
-    auto operator()(const vast::list& xs) {
-      return format_to(out_, "[{}]", ::fmt::join(xs, ", "));
-    }
-    auto operator()(const vast::view<vast::list>& xs) {
-      return format_to(out_, "[{}]", ::fmt::join(xs, ", "));
-    }
-    auto operator()(const vast::map& xs) {
-      return format_to(out_, "{{{}}}", ::fmt::join(xs, ", "));
-    }
-    auto operator()(const vast::view<vast::map>& xs) {
-      return format_to(out_, "{{{}}}", ::fmt::join(xs, ", "));
-    }
-    auto operator()(const vast::record& xs) {
-      return format_to(out_, "<{}>", ::fmt::join(xs, ", "));
-    }
-    auto operator()(const vast::view<vast::record>& xs) {
-      return format_to(out_, "<{}>", ::fmt::join(xs, ", "));
-    }
-
-    template <class T>
-    auto operator()(const T& x) {
-      return format_to(out_, "{}", x);
-    }
-  };
-
-  template <typename Output>
-  struct json_visitor_base {
-    Output out_;
-    json_visitor_base(Output out) : out_{std::move(out)} {
-    }
-
-    auto operator()(caf::none_t) {
-      return format_to(out_, "null");
-    }
-    auto operator()(bool b) {
-      return format_to(out_, b ? "true" : "false");
-    }
-    auto operator()(vast::duration d) {
-      return format_to(out_, "\"{}\"",
-                       vast::detail::fmt_wrapped<vast::duration>{d});
-    }
-    auto operator()(const vast::time& t) {
-      return format_to(out_, "\"{}\"",
-                       vast::detail::fmt_wrapped<vast::time>{t});
-    }
-    auto operator()(const std::string& s) {
-      return (*this)(json_escape_string{s});
-    }
-
-    template <class T>
-    auto operator()(const T& x) {
-      return format_to(out_, "{}", x);
-    }
-  };
-
-  template <typename Output, class PrintTraits>
-  struct json_visitor : json_visitor_base<Output> {
-    using super = json_visitor_base<Output>;
-    PrintTraits print_traits_;
-    json_visitor(Output out, PrintTraits print_traits)
-      : super{out}, print_traits_{std::move(print_traits)} {
-    }
-
-    template <class T>
-    auto operator()(const T& x) {
-      return static_cast<super&>(*this)(x);
-    }
-
-    template <class L>
-    auto format_list(const L& xs) {
-      *this->out_++ = '[';
-      if (!xs.empty()) {
-        print_traits_.inc_indent();
-        bool is_first = true;
-        for (const auto& x : xs) {
-          if (is_first) {
-            is_first = false;
-            print_traits_.format_indent_before_first_item(this->out_);
-          } else {
-            *this->out_++ = ',';
-            print_traits_.format_indent(this->out_);
-          }
-          this->out_ = caf::visit(*this, x);
-        }
-        print_traits_.dec_indent();
-        print_traits_.format_indent_after_last_item(this->out_);
-      }
-      *this->out_ = ']';
-      return this->out_;
-    }
-    auto operator()(const vast::list& xs) {
-      return format_list(xs);
-    }
-    auto operator()(const vast::view<vast::list>& xs) {
-      return format_list(xs);
-    }
-
-    template <class M>
-    auto format_map(const M& xs) {
-      *this->out_++ = '[';
-      if (!xs.empty()) {
-        print_traits_.inc_indent();
-        bool is_first = true;
-        for (const auto& x : xs) {
-          if (is_first) {
-            is_first = false;
-            print_traits_.format_indent_before_first_item(this->out_);
-          } else {
-            *this->out_++ = ',';
-            print_traits_.format_indent(this->out_);
-          }
-          *this->out_++ = '{';
-          print_traits_.inc_indent();
-          print_traits_.format_indent_before_first_item(this->out_);
-
-          print_traits_.format_field_start(this->out_, "key");
-          this->out_ = caf::visit(*this, x.first);
-          *this->out_++ = ',';
-          print_traits_.format_indent(this->out_);
-          print_traits_.format_field_start(this->out_, "value");
-          this->out_ = caf::visit(*this, x.second);
-
-          print_traits_.dec_indent();
-          print_traits_.format_indent_after_last_item(this->out_);
-          *this->out_++ = '}';
-        }
-        print_traits_.dec_indent();
-        print_traits_.format_indent_after_last_item(this->out_);
-      }
-      *this->out_ = ']';
-      return this->out_;
-    }
-    auto operator()(const vast::map& xs) {
-      return format_map(xs);
-    }
-    auto operator()(const vast::view<vast::map>& xs) {
-      return format_map(xs);
-    }
-
-    template <class R>
-    auto format_record(const R& xs) {
-      *this->out_++ = '{';
-      if (!xs.empty()) {
-        print_traits_.inc_indent();
-        bool is_first = true;
-        for (const auto& x : xs) {
-          if (is_first) {
-            is_first = false;
-            print_traits_.format_indent_before_first_item(this->out_);
-          } else {
-            *this->out_++ = ',';
-            print_traits_.format_indent(this->out_);
-          }
-          print_traits_.format_field_start(this->out_, x.first);
-          this->out_ = caf::visit(*this, x.second);
-        }
-        print_traits_.dec_indent();
-        print_traits_.format_indent_after_last_item(this->out_);
-      }
-      *this->out_ = '}';
-      return this->out_;
-    }
-    auto operator()(const vast::record& xs) {
-      return format_record(xs);
-    }
-    auto operator()(const vast::view<vast::record>& xs) {
-      return format_record(xs);
-    }
-  };
-
-  template <bool RemoveSpaces>
-  struct ndjson_print_traits {
-    template <class... Ts>
-    constexpr void dec_indent(Ts&&...) const noexcept {
-    }
-    template <class... Ts>
-    constexpr void inc_indent(Ts&&...) const noexcept {
-    }
-    template <class Output>
-    constexpr void format_indent_before_first_item(Output&) const noexcept {
-    }
-    template <class Output>
-    constexpr void format_indent_after_last_item(Output&) const noexcept {
-    }
-    template <class Output>
-    constexpr void format_indent(Output& out) const {
-      if constexpr (!RemoveSpaces)
-        *out++ = ' ';
-    }
-    template <class Output>
-    void format_field_start(Output& out, std::string_view name) {
-      out = format_to(out,
-                      RemoveSpaces ? "{}:" : "{}: ", json_escape_string{name});
-    }
-  };
-
-  struct json_print_traits {
-    int indent_size;
-    int current_indent{0};
-
-    template <class... Ts>
-    constexpr void dec_indent(Ts&&...) noexcept {
-      --current_indent;
-    }
-    template <class... Ts>
-    constexpr void inc_indent(Ts&&...) noexcept {
-      ++current_indent;
-    }
-    template <class Output>
-    constexpr void format_indent_before_first_item(Output& out) const {
-      format_indent(out);
-    }
-    template <class Output>
-    constexpr void format_indent_after_last_item(Output& out) const {
-      format_indent(out);
-    }
-    template <class Output>
-    constexpr void format_indent(Output& out) const {
-      out = format_to(out, "\n{:<{}}", "", current_indent * indent_size);
-    }
-    template <class Output>
-    void format_field_start(Output& out, std::string_view name) {
-      out = format_to(out, "{}: ", json_escape_string{name});
-    }
-  };
-
-  template <class Data, class FormatContext>
-  auto format(const Data& x, FormatContext& ctx) const {
-    auto do_format = [&x](auto f) { return caf::visit(f, x); };
-    if (presentation == 'a') {
-      return do_format(ascii_visitor{ctx.out()});
-    } else if (presentation == 'j') {
-      if (ndjson) {
-        if (remove_spaces)
-          return do_format(
-            json_visitor{ctx.out(), ndjson_print_traits<true>{}});
-        else
-          return do_format(
-            json_visitor{ctx.out(), ndjson_print_traits<false>{}});
-      } else
-        return do_format(json_visitor{ctx.out(), json_print_traits{indent}});
-    } else if (presentation == 'y') {
-    }
-    return ctx.out();
-  }
-};
-
-template <>
-struct formatter<vast::view<vast::data>> : formatter<vast::data> {};
 
 } // namespace fmt
